requirements.txt
langchain
google-generativeai
python-dotenv
faiss-cpu
langchain_google_genai
langchain_community
langchain-text-splitters
langchain-google-genai
python-dotenv
faiss-cpu
PyPDF2
pdf2image
flask
flask-cors
requests
pdfplumber
jinja2
pdfkit
python-telegram-bot
json_repair
reportlab
